Update TraceKit's regex for stacktraces in Chrome
The stacktrace regex for Chrome doesn't match anonymous functions, so
all anonymous functions in the stacktrace will be removed before
reporting to Sentry.  This causes errors to be attributed to the wrong
files, and is generally extremely confusing.  This change updates the
regex to not *require* a function name be present in the stacktrace.